+++ packages/webpack-plugin/src/loader.ts
@@ -5,7 +5,6 @@ import {
   IdentifierOption,
   processVanillaFile,
   addFileScope,
-  PackageInfo,
   serializeCss,
 } from '@vanilla-extract/integration';
 
@@ -27,7 +26,6 @@ const emptyCssExtractionFile = require.resolve(
 interface LoaderOptions {
   outputCss: boolean;
   identifiers?: IdentifierOption;
-  packageInfo: PackageInfo;
 }
 
 interface InternalLoaderOptions extends LoaderOptions {
@@ -36,13 +34,12 @@ interface InternalLoaderOptions extends LoaderOptions {
 
 export default function (this: LoaderContext, source: string) {
   this.cacheable(true);
-  const { packageInfo } = loaderUtils.getOptions(this) as InternalLoaderOptions;
 
   return addFileScope({
     source,
     filePath: this.resourcePath,
-    packageInfo,
-  }).source;
+    rootPath: this.rootContext,
+  });
 }
 
 export function pitch(this: LoaderContext) {